Release 27.5

+++ blender/source/blender/python/bmesh/bmesh_py_types_customdata.c
@@ -740,8 +740,7 @@ static PyObject *bpy_bmlayercollection_subscript_str(BPy_BMLayerCollection *self
   return NULL;
 }
 
-static PyObject *bpy_bmlayercollection_subscript_int(BPy_BMLayerCollection *self,
-                                                     Py_ssize_t keynum)
+static PyObject *bpy_bmlayercollection_subscript_int(BPy_BMLayerCollection *self, int keynum)
 {
   Py_ssize_t len;
   BPY_BM_CHECK_OBJ(self);
