+++ blender/source/blender/python/bmesh/bmesh_py_types_meshdata.c
@@ -298,7 +298,7 @@ static int mathutils_bmloopcol_set(BaseMathObject *bmo, int UNUSED(subtype))
 
 static int mathutils_bmloopcol_get_index(BaseMathObject *bmo, int subtype, int UNUSED(index))
 {
-  /* Lazy, avoid repeating the case statement. */
+  /* lazy, avoid repeteing the case statement */
   if (mathutils_bmloopcol_get(bmo, subtype) == -1) {
     return -1;
   }
@@ -309,7 +309,7 @@ static int mathutils_bmloopcol_set_index(BaseMathObject *bmo, int subtype, int i
 {
   const float f = bmo->data[index];
 
-  /* Lazy, avoid repeating the case statement. */
+  /* lazy, avoid repeteing the case statement */
   if (mathutils_bmloopcol_get(bmo, subtype) == -1) {
     return -1;
   }
@@ -391,7 +391,7 @@ typedef struct BPy_BMDeformVert {
 /* Mapping Protocols
  * ================= */
 
-static Py_ssize_t bpy_bmdeformvert_len(BPy_BMDeformVert *self)
+static int bpy_bmdeformvert_len(BPy_BMDeformVert *self)
 {
   return self->data->totweight;
 }
